+++ cli/cli/helpers/output_printers/kurtosis_instruction_printer_test.go
@@ -3,13 +3,15 @@ package output_printers
 import (
 	"github.com/kurtosis-tech/kurtosis/api/golang/core/kurtosis_core_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/binding_constructors"
+	"github.com/kurtosis-tech/kurtosis/cli/cli/command_args/run"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
 
-func TestFormatInstruction(t *testing.T) {
-	instruction := binding_constructors.NewKurtosisInstruction(
-		binding_constructors.NewKurtosisInstructionPosition("dummyFile", 12, 4),
+func testInstruction() *kurtosis_core_rpc_api_bindings.KurtosisInstruction {
+	position := binding_constructors.NewKurtosisInstructionPosition("dummyFile", 12, 4)
+	instructionWithNoResult := binding_constructors.NewKurtosisInstruction(
+		position,
 		"my_instruction",
 		`my_instruction("foo", ["bar", "doo"], kwarg1="serviceA", kwarg2=struct(bonjour=42, hello="world"))`,
 		// TODO(gb): for now result is appended manually in the exec command code. This is change once we start doing streaming where the result is displayed right after the instruction code
@@ -19,7 +21,13 @@ func TestFormatInstruction(t *testing.T) {
 			binding_constructors.NewKurtosisInstructionKwarg("serviceA", "kwarg1", true),
 			binding_constructors.NewKurtosisInstructionKwarg(`struct(bonjour=42, hello="world")`, "kwarg2", false),
 		})
-	formattedInstruction := formatInstruction(instruction)
+	instructionResult := "RESULT"
+	return binding_constructors.AddResultToKurtosisInstruction(instructionWithNoResult, &instructionResult)
+}
+
+func TestFormatInstruction_Executable(t *testing.T) {
+	instruction := testInstruction()
+	formattedInstruction := formatInstruction(instruction, run.Executable)
 	expectedResult := `# from dummyFile[12:4]
 my_instruction(
     "foo",
@@ -32,7 +40,28 @@ my_instruction(
         bonjour = 42,
         hello = "world",
     ),
-)`
+)
+RESULT`
+	require.Equal(t, expectedResult, formattedInstruction)
+}
+
+func TestFormatInstruction_Detailed(t *testing.T) {
+	instruction := testInstruction()
+	formattedInstruction := formatInstruction(instruction, run.Detailed)
+	expectedResult := `> my_instruction
+> 	foo
+> 	["bar", "doo"]
+> 	kwarg1=serviceA
+> 	kwarg2=struct(bonjour=42, hello="world")
+RESULT`
+	require.Equal(t, expectedResult, formattedInstruction)
+}
+
+func TestFormatInstruction_Brief(t *testing.T) {
+	instruction := testInstruction()
+	formattedInstruction := formatInstruction(instruction, run.Brief)
+	expectedResult := `> my_instruction foo kwarg1=serviceA
+RESULT`
 	require.Equal(t, expectedResult, formattedInstruction)
 }
 
@@ -43,7 +72,7 @@ func TestFormatInstruction_FormattingFail(t *testing.T) {
 		// This has issues with the quotes not being escaped
 		`print("UNSUPPORTED_TYPE['ModuleOutput(grafana_info=GrafanaInfo(dashboard_path="/d/QdTOwy-nz/eth2-merge-kurtosis-module-dashboard?orgId=1", user="admin", password="admin"))']")`,
 		[]*kurtosis_core_rpc_api_bindings.KurtosisInstructionArg{})
-	formattedInstruction := formatInstruction(instruction)
+	formattedInstruction := formatInstruction(instruction, run.Executable)
 	// failure to format -> the instruction is returned with no formatting applied
 	expectedResult := `# from dummyFile[12:4]
 print("UNSUPPORTED_TYPE['ModuleOutput(grafana_info=GrafanaInfo(dashboard_path="/d/QdTOwy-nz/eth2-merge-kurtosis-module-dashboard?orgId=1", user="admin", password="admin"))']")`