+++ core/server/api_container/server/startosis_engine/startosis_const.go
@@ -1,17 +1,10 @@
 package startosis_engine
 
 const (
-	MainFileName  = "main.star"
-	TypesFileName = "types.proto"
+	MainFileName = "main.star"
 
-	ModuleInputTypeName  = "ModuleInput"
-	ModuleOutputTypeName = "ModuleOutput"
-
-	MainInputArgName = "input_args"
+	MainInputArgName = "args"
 	EmptyInputArgs   = "{}" // empty JSON
 
 	ModuleIdPlaceholderForStandaloneScripts = "DEFAULT_MODULE_ID_FOR_SCRIPT"
-
-	Representative    = true
-	NotRepresentative = false
 )