+++ internal_testsuites/golang/testsuite/startosis_module_test/shared_constants_test.go
@@ -4,10 +4,4 @@ const (
 	isPartitioningEnabled = false
 	emptyExecuteParams    = "{}"
 	defaultDryRun         = false
-
-	validModuleNoTypeTestName = "valid-module-no-type"
-	validModuleNoTypeRelPath  = "../../../startosis/valid-kurtosis-module-no-type"
-
-	validModuleNoModuleInputTypeTestName = "valid-module-no-input-type"
-	validModuleNoModuleInputTypeRelPath  = "../../../startosis/valid-kurtosis-module-no-module-input-type"
 )