+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ### Changes
 - Re-activate tests that had to be skipped because of the "Remove support for protobuf in Startosis" breaking change
+- Renamed `input_args` to `args`. All Starlark packages should update `run(input_args)` to `run(args)`
 
 # 0.55.0